send-read-emails: refactor client code
