+++ packages/send-and-read-emails/client/react/src/styles.js
@@ -1,8 +1,15 @@
 export const styles = {
-  App: {
+  EmailClient: {
+    contentContainer: {
+      display: 'flex',
+      flexDirection: 'row',
+      width: '80vw',
+      maxHeight: '50vh',
+      margin: '10vh auto',
+    },
     statusBar: {
       width: '80vw',
-      margin: '0 auto',
+      margin: '6em auto 0',
       height: '60px',
       border: '2px solid #4169e1',
       background: 'rgba(0,42,245,0.8)',
@@ -12,17 +19,10 @@ export const styles = {
       alignItems: 'center',
     },
     statusBarText: {
-      padding: 0,
+      padding: '1em',
       marginBlock: 0,
       fontWeight: 'bold',
       color: 'white',
-    },
-    contentContainer: {
-      display: 'flex',
-      flexDirection: 'row',
-      width: '80vw',
-      maxHeight: '50vh',
-      margin: '10vh auto',
     },
   },
   ReadEmails: {
